Extract categorisation to shared transform
